fix: minor change

+++ pintos/include/threads/thread.h
@@ -114,7 +114,7 @@ struct  thread {
 	struct list_elem elem;              /* 리스트 원소. */
 	struct list_elem allelem;           /* 전체 리스트용 원소 */
 
-	uint64_t exit_status;               /* 종료 코드 */
+	int exit_status;               /* 종료 코드 */
 
 #ifdef USERPROG
 	/* `userprog/process.c`가 관리한다. */
